backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy 
pymysql
pydantic
pydantic-core
pydantic-settings
typing-extensions
databases 
mysql-connector-python
python-dotenv
sqlmodel
pyjwt
passlib
alembic
pandas
aiofiles
cryptography
redis
python-multipart

pillow
requests
ftfy 
regex 
tqdm
open_clip_torch
# git+https://github.com/openai/CLIP.git
pymilvus
transformers

numpy==1.24.4
elasticsearch
elasticsearch-dsl
